# refactor(gridintersect): faster __init__ and better shapely 2.0 compat