--- a/etc/environment.yml
+++ b/etc/environment.yml
@@ -36,7 +36,7 @@ dependencies:
   - fiona
   - descartes
   - pyproj
-  - shapely
+  - shapely>=1.8
   - geos
   - geojson
   - vtk