--- a/setup.cfg
+++ b/setup.cfg
@@ -69,7 +69,7 @@ optional =
     rasterio; platform_system!='Windows'
     rasterstats; platform_system!='Windows'
     scipy
-    shapely
+    shapely >=1.8
     vtk
     xmipy
 doc =