This PR does the following:

- enforce minimum shapely 1.8, also check that it works with shapely 2.0 (beta 1 pre-release was this week; [new docs here](https://shapely.readthedocs.io/en/latest/))
- change types `geoms` and `cellids` from lists to numpy arrays, which are are faster to process; but keep list return types for deprecated functions
- corrections to docstrings
- focus on speed-up for `_rect_grid_to_geoms_cellids` (called from `__init__`), building arrays of coordinates for constructing grid cell polygons
  - before this PR, testing with a 500x500 grid, `GridIntersect(sgr, method="vertex", rtree=False)` (i.e. calling `__init__`) would take 3.09 s ± 61.2 ms
  - with this PR and shapely 1.8.5, 2.3 s ± 79.5 ms, so 1.3x faster
  - with this PR and shapely 2.0b1, 146 ms ± 184 µs, so about 15x faster
- fix-up `filter_query_result` so it works with shapely 2.0, but this performs similar to shapely 1.8

There are probably some performance gains to be made with `ix.intersect(shp)`, but they haven't been exposed yet. Shapely 2.0 users can also do array-based calculations like `shapely.area(shapely.intersection(ix.geoms, shp)).sum()`.