@dbrakenhoff 2022-10-18
Also curious if the steps (1 query, 2 filter, 3 intersect) still make sense when using the array-based solutions, but that's for another PR.