@mwtoews 2022-10-19
@dbrakenhoff other array-based improvements can be added later, and I would expect some performance gains. The module should be able to handle both shapely 1.8 and 2+ versions using `if SHAPELY_GE_20` blocks, as done in this PR. Essentially for-loops should be eliminated where possible. Shapely 2.0 can still use similar approaches (query, filter, intersect), but you may find the speed of things like `shapely.intersects(ix.geoms, shp)` very fast, even with large grids.

For anyone that wants to test shapely 2.0b1, [here are some tips](https://github.com/shapely/shapely/discussions/1578).